# docs/normative/v0.6/contracts/AI-05_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Paquetería Local Culiacán API
  version: 0.6.0
  description: Contrato REST normativo v0.6. REST/PostgreSQL son autoridad; SignalR distribuye eventos confirmados. Seguridad
    tenant, tracking público y outbox siguen contratos ejecutables.
servers:
- url: /api/v1
security:
- oidc: []
...
paths:
...
  /routes:
    post:
      tags:
      - Routing
      operationId: createRoute
      description: >-
        Creates a tenant-owned manual route for an OWN, ACTIVE driver who is
        currently eligible under DSP-001 in the active OrganizationContext.
        City and optional service area must be coherent with that driver. The
        route starts DRAFT at version 1. Every successful mutation is audited
        and emits RouteChanged only through the transactional outbox after
        commit. Reusing the same Idempotency-Key and canonical request returns
        the stored 201 without creating another route; an idempotency mismatch
        returns 409. EXTERNAL and ALLY_CAPACITY routing are not enabled by
        RTE-001.
      x-required-capability: Dispatch
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteRequest'
      responses:
        '201':
          description: DRAFT route created at version 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/RoutingConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      tags:
      - Routing
      operationId: listRoutes
      description: >-
        Returns only routes visible in the active OrganizationContext. The
        server owns the bounded page size; clients cannot configure it. Access
        follows the existing Operations read capability.
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/RouteStatus'
      - name: driver_id
        in: query
        schema:
          type: string
          format: uuid
      - name: scheduled_for
        in: query
        schema:
          type: string
          format: date
      - name: cursor
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Cursor-paginated tenant routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /routes/{routeId}:
    get:
      tags:
      - Routing
      operationId: getRoute
      description: >-
        Returns the authoritative route and every current stop ordered by
        sequence ascending. Clients reconstruct state from this resource on
        load, refresh, reconnect and after RouteChanged. Access follows the
        existing Operations read capability.
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/RouteId'
      responses:
        '200':
          description: Authoritative route detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /routes/{routeId}/stops:
    post:
      tags:
      - Routing
      operationId: addRouteStop
      description: >-
        Adds one DELIVERY stop to a DRAFT route and links the existing OWN
        ACCEPTED or ACTIVE assignment for an order visible in the active tenant
        to that route in the same transaction.
        The route driver must be OWN and ACTIVE; assignment.driver_id must equal
        route.driver_id; the order and driver must remain compatible under
        DSP-001. A stable PostgreSQL serialization decision by order_id prevents
        duplicates and membership in another DRAFT, PLANNED or ACTIVE route.
        The inserted sequence is the next contiguous integer and route.version
        increments exactly once. Audit and RouteChanged outbox records commit
        atomically. The client cannot supply stop_type. Reusing the same
        Idempotency-Key and canonical request returns the stored success without
        another mutation; a different request for the key returns 409.
      x-required-capability: Dispatch
      x-rte-001-assignment-types:
      - OWN
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/RouteId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRouteStopRequest'
      responses:
        '201':
          description: Stop and assignment route link created atomically
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/RoutingConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /routes/{routeId}/stops/{stopId}:
    delete:
      tags:
      - Routing
      operationId: removeRouteStop
      description: >-
        Removes a stop only from a DRAFT route after checking expected_version.
        In one transaction it clears route_id only on the linked assignment,
        compacts the remaining sequence to 1..N, increments route.version once,
        and writes audit plus RouteChanged outbox evidence. Reusing the same
        Idempotency-Key and canonical request is safe and returns the stored
        success; an idempotency mismatch returns 409.
      x-required-capability: Dispatch
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/RouteId'
      - $ref: '#/components/parameters/RouteStopId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - name: expected_version
        in: query
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Updated route after atomic stop removal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/RoutingConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /routes/{routeId}/stops/order:
    put:
      tags:
      - Routing
      operationId: reorderRouteStops
      description: >-
        Reorders a DRAFT route under a route lock. stop_ids must contain every
        current stop exactly once, with no duplicate, foreign or missing ID.
        The server writes contiguous sequence values 1..N and increments the
        route version exactly once in the same transaction as audit and the
        RouteChanged outbox record. A stale expected_version returns 409. Of two
        concurrent reorders using the same version exactly one succeeds and the
        other receives 409. Idempotent replay returns the stored success without
        another version increment.
      x-required-capability: Dispatch
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/RouteId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderRouteStopsRequest'
      responses:
        '200':
          description: Route stops reordered atomically
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/RoutingConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    RouteId:
      name: routeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RouteStopId:
      name: stopId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    OrganizationContext:
      name: X-Organization-Id
      in: header
      required: true
      description: Organización activa; debe corresponder a una membresía ACTIVE del usuario.
      schema:
        type: string
        format: uuid
  responses:
...
    RoutingConflict:
      description: >-
        RTE-001 stale version, duplicate route membership, active-route
        conflict, route or assignment state conflict, driver/order
        incompatibility, eligibility/capacity rejection or idempotency mismatch.
        Internal evidence and cross-tenant existence are never exposed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated actor lacks capability inside an otherwise visible tenant context
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UniformNotFound:
      description: Missing, cross-tenant, expired or revoked resource. Response must not reveal existence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ServiceUnavailable:
      description: >-
        Required provider or infrastructure dependency is unavailable. Every
        authenticated operation can report it while the identity context
        cannot be resolved, and every tenant operation while the mandatory
        PLATFORM_ADMIN tenant-activation audit cannot be written
        (AI05-DECLARE-EMITTED-ERRORS).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    RouteStatus:
      type: string
      enum:
      - DRAFT
      - PLANNED
      - ACTIVE
      - COMPLETED
      - CANCELLED
    CreateRouteRequest:
      type: object
      additionalProperties: false
      required:
      - driver_id
      - city_id
      properties:
        driver_id:
          type: string
          format: uuid
        city_id:
          type: string
          format: uuid
        service_area_id:
          type:
          - string
          - 'null'
          format: uuid
        scheduled_for:
          type:
          - string
          - 'null'
          format: date
    AddRouteStopRequest:
      type: object
      additionalProperties: false
      required:
      - order_id
      - expected_version
      properties:
        order_id:
          type: string
          format: uuid
        expected_version:
          type: integer
          minimum: 1
    ReorderRouteStopsRequest:
      type: object
      additionalProperties: false
      required:
      - expected_version
      - stop_ids
      properties:
        expected_version:
          type: integer
          minimum: 1
        stop_ids:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
          description: Must contain every current route stop exactly once.
...
    Route:
      type: object
      required:
      - id
      - status
      - version
      - driver_id
      - city_id
      - service_area_id
      - scheduled_for
      - assignment_cost_cents_total
      - stop_count
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/RouteStatus'
        version:
          type: integer
          minimum: 1
        driver_id:
          type: string
          format: uuid
        city_id:
          type: string
          format: uuid
        service_area_id:
          type:
          - string
          - 'null'
          format: uuid
        scheduled_for:
          type:
          - string
          - 'null'
          format: date
        assignment_cost_cents_total:
          type: integer
          format: int64
          minimum: 0
          description: >-
            Server-side SUM of cost_cents for ACCEPTED or ACTIVE assignments
            currently linked to the route. The total is calculated, not stored.
        stop_count:
          type: integer
          minimum: 0
    RouteDetail:
      allOf:
      - $ref: '#/components/schemas/Route'
      - type: object
        required:
        - stops
        properties:
          stops:
            type: array
            description: Current route stops ordered by sequence ascending.
            items:
              $ref: '#/components/schemas/RouteStop'
    RoutePage:
      type: object
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Route'
        next_cursor:
          type:
          - string
          - 'null'
...